Add basic AnnData metadata header CSFV
